pyproject.toml
# poetry run pip install git+https://github.com/openai/whisper.git@v20231117
# manually installed through pip due to issue building wheel in poetry

[tool.poetry]
name = ""
version = "0.1.0"
description = "Multi-modal document analysis system with AI capabilities"
authors = ["Sainath Dey"]
readme = "README.md"
packages = [
    { include = "app", from = "." },
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
dotenv = "^0.9.9"
asyncio = "^3.4.3"
llama-index = "^0.12.25"
pymupdf4llm = "^0.0.17"
langchain = "^0.3.21"
pdf2docx = "^0.5.8"
llama-index-embeddings-ollama = "^0.6.0"
ffmpeg-python = "^0.2.0"
numpy = "2.1"
qdrant-client = "^1.13.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.5"
docx2txt = "^0.8"
googlesearch-python = "^1.3.0"
langchain-ollama = "^0.3.0"
streamlit = "^1.43.2"
pytest-asyncio = "^0.25.3"
groq = "^0.20.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
crawl4ai = "^0.5.0.post8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

